% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_words.R
\name{count_words}
\alias{count_words}
\title{Count words in child and adult speech}
\usage{
count_words(this.transcript, use.mor = TRUE, POS_regex = NULL,
  mor = "mor_word", orth = "orth_word", debug = FALSE)
}
\description{
Count words in child and adult speech
}

