% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_mor_word.R
\name{parse_mor_word}
\alias{parse_mor_word}
\title{Extract information from %mor tier words in CHAT transcripts}
\usage{
parse_mor_word(mor_word, extract = c("POS", "word", "gloss"))
}
\description{
Extract information from %mor tier words in CHAT transcripts
}

